BOAC-841, if user is ASC then search excludes 'inactive' students
